requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
email-validator
python-dotenv
requests
bcrypt
pandas
openpyxl
jinja2
markupsafe
pywebview
pyinstaller
Pillow
